+++ examples/rpc/rpc-server.cpp
@@ -7,19 +7,8 @@
 #endif
 
 #include "ggml-rpc.h"
-#include <memory>
 #include <string>
-#ifndef _WIN32
-#  include <sys/socket.h>
-#  include <sys/types.h>
-#  include <arpa/inet.h>
-#  include <netinet/in.h>
-#  include <netinet/tcp.h>
-#  include <netdb.h>
-#  include <unistd.h>
-#endif
 #include <stdio.h>
-#include <stdlib.h>
 
 static ggml_backend_t create_backend() {
     ggml_backend_t backend = NULL;
@@ -55,91 +44,27 @@ static void get_backend_memory(size_t * free_mem, size_t * total_mem) {
 #endif
 }
 
-static std::shared_ptr<socket_t> make_socket(sockfd_t fd) {
-#ifdef _WIN32
-    if (fd == INVALID_SOCKET) {
-        return nullptr;
-    }
-#else
-    if (fd < 0) {
-        return nullptr;
-    }
-#endif
-    return std::make_shared<socket_t>(fd);
-}
-
-static std::shared_ptr<socket_t> create_server_socket(const char * host, int port) {
-    auto sockfd = socket(AF_INET, SOCK_STREAM, 0);
-    auto sock = make_socket(sockfd);
-    if (sock == nullptr) {
-        return nullptr;
-    }
-
-    struct sockaddr_in serv_addr;
-    serv_addr.sin_family = AF_INET;
-    serv_addr.sin_addr.s_addr = inet_addr(host);
-    serv_addr.sin_port = htons(port);
-
-    if (bind(sockfd, (struct sockaddr *) &serv_addr, sizeof(serv_addr)) < 0) {
-        return nullptr;
-    }
-    if (listen(sockfd, 5) < 0) {
-        return nullptr;
-    }
-    return sock;
-}
-
 int main(int argc, char * argv[])
 {
-#ifdef _WIN32
-    WSADATA wsaData;
-    int res = WSAStartup(MAKEWORD(2, 2), &wsaData);
-    if (res != 0) {
-        fprintf(stderr, "WSAStartup failed: %d\n", res);
-        return 1;
-    }
-#endif
     if (argc < 3) {
         fprintf(stderr, "Usage: %s <host> <port>\n", argv[0]);
         return 1;
     }
     const char * host = argv[1];
     int port = std::stoi(argv[2]);
-
+    if (port <= 0 || port > 65535) {
+        fprintf(stderr, "Invalid port number: %d\n", port);
+        return 1;
+    }
     ggml_backend_t backend = create_backend();
     if (!backend) {
         fprintf(stderr, "Failed to create backend\n");
         return 1;
     }
-
     printf("Starting RPC server on %s:%d\n", host, port);
-    auto server_socket = create_server_socket(host, port);
-    if (server_socket == nullptr) {
-        fprintf(stderr, "Failed to create server socket\n");
-        return 1;
-    }
-    while (true) {
-        auto client_socket_fd = accept(server_socket->fd, NULL, NULL);
-        auto client_socket = make_socket(client_socket_fd);
-        if (client_socket == nullptr) {
-            fprintf(stderr, "Failed to accept client connection\n");
-            return 1;
-        }
-        // set TCP_NODELAY to disable Nagle's algorithm
-        int flag = 1;
-        int ret = setsockopt(client_socket->fd, IPPROTO_TCP, TCP_NODELAY, (char *) &flag, sizeof(int));
-        if (ret < 0) {
-            fprintf(stderr, "Failed to set TCP_NODELAY\n");
-            continue;
-        }
-        size_t free_mem, total_mem;
-        get_backend_memory(&free_mem, &total_mem);
-        printf("Accepted client connection, free_mem=%zu, total_mem=%zu\n", free_mem, total_mem);
-        rpc_serve_client(backend, client_socket->fd, free_mem, total_mem);
-        printf("Client connection closed\n");
-    }
-#ifdef _WIN32
-    WSACleanup();
-#endif
+    size_t free_mem, total_mem;
+    get_backend_memory(&free_mem, &total_mem);
+    std::string endpoint = std::string(host) + ":" + std::to_string(port);
+    start_rpc_server(backend, endpoint.c_str(), free_mem, total_mem);
     return 0;
 }
